pyproject.toml
[project]
name = "pytriton-serve-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "nvidia-pytriton>=0.7.0",
    "torch",
]
